moved all functions to one location
