+++ tests/function-invoke/function-loader.test.js
@@ -3,10 +3,10 @@ var funcLoader = require('../../src/function-loader');
 
 test('function loader tests', function (group) {
     group.test('should load function.json and extract function', function (t) {
-        const actual = funcLoader.loadFunction('simple-test-func', 'tests/function-invoke');
+        const actual = funcLoader.loadFunction('simple-test-func', 'tests/test-functions');
         const expected = {
             config: {},
-            function: require('./simple-test-func/index'),
+            function: require('../test-functions/simple-test-func/index'),
             sampleData:{}
         };
 
@@ -15,12 +15,12 @@ test('function loader tests', function (group) {
     });
 
     group.test('should load function.json and single exported function', function (t) {
-        const actual = funcLoader.loadFunction('single-export-func', 'tests/function-invoke');
+        const actual = funcLoader.loadFunction('single-export-func', 'tests/test-functions');
         const expected = {
             config: {
                 "disabled": false
             },
-            function: require('./single-export-func/').mainFunction,
+            function: require('../test-functions/single-export-func').mainFunction,
             sampleData:{}
         };
 
@@ -29,12 +29,12 @@ test('function loader tests', function (group) {
     });
 
     group.test('should load function.json and run function', function (t) {
-        const actual = funcLoader.loadFunction('run-property-func', 'tests/function-invoke');
+        const actual = funcLoader.loadFunction('run-property-func', 'tests/test-functions');
         const expected = {
             config: {
                 "disabled": false
             },
-            function: require('./run-property-func').run,
+            function: require('../test-functions/run-property-func').run,
             sampleData:{}
         }
 
@@ -43,13 +43,13 @@ test('function loader tests', function (group) {
     });
 
     group.test('should return if function.json has entryPoint Configured', function (t) {
-        const actual = funcLoader.loadFunction('entrypoint-func', 'tests/function-invoke');
+        const actual = funcLoader.loadFunction('entrypoint-func', 'tests/test-functions');
         const expected = {
             config: {
                 "entryPoint": "mainFunction",
                 "disabled": false
             },
-            function: require('./entrypoint-func').mainFunction,
+            function: require('../test-functions/entrypoint-func').mainFunction,
             sampleData:{}
         };
 
@@ -58,10 +58,10 @@ test('function loader tests', function (group) {
     });
 
     group.test('should load sample.dat if there', function (t) {
-        const actual = funcLoader.loadFunction('sampledata-func', 'tests/function-invoke');
+        const actual = funcLoader.loadFunction('sampledata-func', 'tests/test-functions');
         const expected = {
             config: {},
-            function: require('./sampledata-func'),
+            function: require('../test-functions/sampledata-func'),
             sampleData: {
                 "sampleData": "001",
                 "somethingelse": "000"
@@ -82,7 +82,7 @@ test('function loader tests', function (group) {
 
     group.test('should throw if no function.json', function (t) {
         t.throws(_ => {
-            funcLoader.loadFunction('no-functionjson-func', 'tests/function-invoke');
+            funcLoader.loadFunction('no-functionjson-func', 'tests/test-functions');
         }, new RegExp('Could not find a function.*no function.json file.'));
 
         t.end();
@@ -90,7 +90,7 @@ test('function loader tests', function (group) {
 
     group.test('should throw if function.json has invalid entryPoint', function (t) {
         t.throws(_ => {
-            funcLoader.loadFunction('entrypoint-invalid-func', 'tests/function-invoke');
+            funcLoader.loadFunction('entrypoint-invalid-func', 'tests/test-functions');
         }, new RegExp('Could not find a function.*failed on the Azure Functions resolution rules.'));
 
         t.end();
@@ -98,7 +98,7 @@ test('function loader tests', function (group) {
 
     group.test('should throw if module does not export function', function (t) {
         t.throws(_ => {
-            funcLoader.loadFunction('multiple-function-export-func', 'tests/function-invoke');
+            funcLoader.loadFunction('multiple-function-export-func', 'tests/test-functions');
         }, new RegExp('Could not find a function.*failed on the Azure Functions resolution rules.'));
 
         t.end();
