--- tests/http-trigger/sampledata-func/index.js
@@ -1,4 +1,0 @@
-module.exports = function(context) {
-    //simplest function there could be
-    context.done();
-}
