--- tests/http-trigger/sampledata-func/sample.dat
@@ -1,4 +1,0 @@
-{
-    "sampleData": "001",
-    "somethingelse": "000"
-}
