+++ tests/queue-trigger/queue-trigger.test.js
@@ -2,7 +2,7 @@ var test = require("tape");
 var func = require('../../src/index.js');
 
 test('Queue trigger', function(group){
-    var queueFunc = func('queue-func', {dirname: __dirname}); 
+    var queueFunc = func('queue-func', {dirname: 'tests/test-functions'}); 
     
     group.test('promise should return output on binding', function (t){
         t.plan(1);
