+++ tests/test-functions/entrypoint-func/function.json
