+++ tests/test-functions/entrypoint-func/index.js
