+++ tests/test-functions/entrypoint-invalid-func/function.json
