+++ tests/test-functions/entrypoint-invalid-func/index.js
