+++ tests/test-functions/multiple-function-export-func/function.json
