+++ tests/test-functions/multiple-function-export-func/index.js
