+++ tests/test-functions/no-functionjson-func/index.js
