+++ tests/test-functions/run-property-func/function.json
