+++ tests/test-functions/run-property-func/index.js
