+++ tests/test-functions/simple-test-func/index.js
