+++ tests/test-functions/single-export-func/function.json
