+++ tests/test-functions/single-export-func/index.js
